finish

+++ app/Http/Controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class ListingController extends Controller
@@ -16,14 +18,14 @@ class ListingController extends Controller
     //     ]);
     // }
 
-        // Show all listings
+    // Show all listings
     public function index() {
         return view('listings.index', [
             'listings' => Listing::latest()->filter
             (request(['tag', 'search']))->paginate(4)
         ]);
 
-        // simplePaginate(2) -> Para paginação sem o número, apenas 'anterior ou proximo'
+    // simplePaginate(2) -> Para paginação sem o número, apenas 'anterior ou proximo'
     }
 
     // Show Create Form
@@ -47,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -61,6 +65,12 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing) {
+
+        // Check listing owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action.');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -80,18 +90,29 @@ class ListingController extends Controller
         return back()->with('message', 'Listing updated successfully!');
     }
 
-        // Delete Listing 
-        public function destroy(Listing $listing) {
-            $listing->delete();
-    
-            return redirect('/')->with('message', 'Listing deleted successfully!');
+    // Delete Listing 
+    public function destroy(Listing $listing) {
+
+        // Check listing owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action.');
         }
+
+        $listing->delete();
+    
+        return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
 
     // Show single listing
     public function show(Listing $listing) {
         return view('listings.show', [
             'listing' => $listing
         ]);
+    }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => User::find(auth()->id())->listings()->get()]);
     }
 
 //     public function show($id) {
